# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payment-details": {
      "get": {
        "description": "Get all payment details records",
        "operationId": "PaymentDetailsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all payment details",
        "tags": [
          "/payment-details"
        ]
      },
      "post": {
        "description": "Create a new payment details record",
        "operationId": "PaymentDetailsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDetailsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create payment details",
        "tags": [
          "/payment-details"
        ]
      }
    },
...
    "/payment-details/{id}": {
      "get": {
        "description": "Get a single payment details record by id",
        "operationId": "PaymentDetailsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get payment details by id",
        "tags": [
          "/payment-details"
        ]
      },
      "patch": {
        "description": "Update an existing payment details record",
        "operationId": "PaymentDetailsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDetailsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update payment details",
        "tags": [
          "/payment-details"
        ]
      },
      "delete": {
        "description": "Delete a payment details record",
        "operationId": "PaymentDetailsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete payment details",
        "tags": [
          "/payment-details"
        ]
      }
    },
    "/payment-details/{id}/activate": {
      "patch": {
        "description": "Sets the given record as active and deactivates all others. Only one record can be active at a time.",
        "operationId": "PaymentDetailsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Activate payment details",
        "tags": [
          "/payment-details"
        ]
      }
    },
...
    "/orders": {
      "post": {
        "description": "Place a new order. Validates stock for each variant, snapshots prices, and can apply a discount coupon by code. User can be authenticated or guest.",
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create order",
        "tags": [
          "/orders"
        ]
      },
      "get": {
        "description": "Paginated list of all orders with optional filters by order_status and payment_status. Admin only.",
        "operationId": "OrderController_findAll",
        "parameters": [
          {
            "name": "order_status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "CONFIRMED",
                "PROCESSING",
                "SHIPPED",
                "DELIVERED",
                "CANCELLED",
                "RETURNED"
              ]
            }
          },
          {
            "name": "payment_status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PAID",
                "FAILED",
                "REFUNDED"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all orders",
        "tags": [
          "/orders"
        ]
      }
    },
    "/orders/me": {
      "get": {
        "description": "Paginated list of orders for the authenticated user with optional filters by order_status and payment_status.",
        "operationId": "OrderController_findMyOrders",
        "parameters": [
          {
            "name": "order_status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "CONFIRMED",
                "PROCESSING",
                "SHIPPED",
                "DELIVERED",
                "CANCELLED",
                "RETURNED"
              ]
            }
          },
          {
            "name": "payment_status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PAID",
                "FAILED",
                "REFUNDED"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get my orders",
        "tags": [
          "/orders"
        ]
      }
    },
    "/orders/me/{id}": {
      "get": {
        "description": "Get full details of a single order that belongs to the authenticated user.",
        "operationId": "OrderController_findMyOrderById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get my order by id",
        "tags": [
          "/orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "description": "Get full order details by MongoDB id. Admin only.",
        "operationId": "OrderController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get order by id",
        "tags": [
          "/orders"
        ]
      },
      "patch": {
        "description": "Update editable order fields (items, customer, delivery, payment method, comment). Recalculates totals if items are changed. Admin only.",
        "operationId": "OrderController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update order",
        "tags": [
          "/orders"
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "description": "Change the fulfillment status of an order. Admin only.",
        "operationId": "OrderController_updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update order status",
        "tags": [
          "/orders"
        ]
      }
    },
    "/orders/{id}/payment-status": {
      "patch": {
        "description": "Change the payment status of an order. Optionally record the payment transaction id. Admin only.",
        "operationId": "OrderController_updatePaymentStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update payment status",
        "tags": [
          "/orders"
        ]
      }
    },
    "/orders/{id}/ttn": {
      "patch": {
        "description": "Attach the Nova Post tracking number (ТТН) to an order. Admin only.",
        "operationId": "OrderController_setTtn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTtnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Set Nova Post TTN",
        "tags": [
          "/orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urban Tab API",
    "description": "API документація",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePaymentDetailsDto": {
        "type": "object",
        "properties": {
          "last_name": {
            "type": "string",
            "example": "Шевченко",
            "description": "Recipient's last name"
          },
          "first_name": {
            "type": "string",
            "example": "Тарас",
            "description": "Recipient's first name"
          },
          "middle_name": {
            "type": "string",
            "example": "Григорович",
            "description": "Recipient's middle name"
          },
          "iban": {
            "type": "string",
            "example": "UA123456789012345678901234567",
            "description": "International Bank Account Number (IBAN)"
          },
          "edrpou": {
            "type": "string",
            "example": "12345678",
            "description": "Ukrainian state registry code (ЄДРПОУ)"
          },
          "bank_name": {
            "type": "string",
            "example": "ПриватБанк",
            "description": "Name of the bank"
          }
        },
        "required": [
          "last_name",
          "first_name",
          "iban",
          "edrpou",
          "bank_name"
        ]
      },
      "UpdatePaymentDetailsDto": {
        "type": "object",
        "properties": {
          "last_name": {
            "type": "string",
            "example": "Шевченко",
            "description": "Recipient's last name"
          },
          "first_name": {
            "type": "string",
            "example": "Тарас",
            "description": "Recipient's first name"
          },
          "middle_name": {
            "type": "string",
            "example": "Григорович",
            "description": "Recipient's middle name"
          },
          "iban": {
            "type": "string",
            "example": "UA123456789012345678901234567",
            "description": "International Bank Account Number (IBAN)"
          },
          "edrpou": {
            "type": "string",
            "example": "12345678",
            "description": "Ukrainian state registry code (ЄДРПОУ)"
          },
          "bank_name": {
            "type": "string",
            "example": "ПриватБанк",
            "description": "Name of the bank"
          }
        }
      },
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "variant_id": {
            "type": "string",
            "example": "664f1b2c3d4e5f6a7b8c9d0e"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1
          }
        },
        "required": [
          "variant_id",
          "quantity"
        ]
      },
      "CreateOrderCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Іван Петренко"
          },
          "phone": {
            "type": "string",
            "example": "+380991234567"
          },
          "email": {
            "type": "string",
            "example": "ivan@example.com"
          }
        },
        "required": [
          "name",
          "phone",
          "email"
        ]
      },
      "CreateOrderAddressDto": {
        "type": "object",
        "properties": {
          "city_name": {
            "type": "string",
            "example": "Київ"
          },
          "warehouse_description": {
            "type": "string",
            "example": "Відділення №1: вул. Хрещатик, 1"
          },
          "warehouse_number": {
            "type": "number",
            "example": 1
          },
          "street": {
            "type": "string",
            "example": "вул. Хрещатик"
          },
          "building": {
            "type": "string",
            "example": "1А"
          },
          "apartment": {
            "type": "string",
            "example": "42"
          }
        },
        "required": [
          "city_name"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          },
          "customer": {
            "$ref": "#/components/schemas/CreateOrderCustomerDto"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "CASH",
              "IBAN",
              "LIQPAY",
              "MONOPAY"
            ],
            "example": "LIQPAY"
          },
          "delivery_method": {
            "type": "string",
            "enum": [
              "NOVA_POST",
              "COURIER",
              "PICKUP"
            ],
            "example": "NOVA_POST"
          },
          "delivery_address": {
            "$ref": "#/components/schemas/CreateOrderAddressDto"
          },
          "comment": {
            "type": "string",
            "example": "Зателефонуйте перед відправкою"
          },
          "coupon_code": {
            "type": "string",
            "example": "AB12CD34EF",
            "description": "User-entered discount code"
          }
        },
        "required": [
          "items",
          "customer",
          "payment_method",
          "delivery_method"
        ]
      },
      "OrderResponseCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Іван Петренко"
          },
          "phone": {
            "type": "string",
            "example": "+380991234567"
          },
          "email": {
            "type": "string",
            "example": "ivan@example.com"
          }
        },
        "required": [
          "name",
          "phone",
          "email"
        ]
      },
      "OrderResponseItemDto": {
        "type": "object",
        "properties": {
          "variant_id": {
            "type": "string",
            "example": "664f1b2c3d4e5f6a7b8c9d0e"
          },
          "product_id": {
            "type": "string",
            "example": "664f1b2c3d4e5f6a7b8c9d0f"
          },
          "name": {
            "type": "string",
            "example": "Стілець кухонний Fillando"
          },
          "sku": {
            "type": "string",
            "example": "FIL-CHAIR-001"
          },
          "vendor_sku": {
            "type": "object",
            "example": "VEND-CH-42",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 1299
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "line_total": {
            "type": "number",
            "example": 2598
          },
          "image": {
            "type": "object",
            "example": "https://cdn.example.com/product/image.webp",
            "nullable": true
          }
        },
        "required": [
          "variant_id",
          "product_id",
          "name",
          "sku",
          "vendor_sku",
          "price",
          "quantity",
          "line_total",
          "image"
        ]
      },
      "OrderResponseDiscountDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "AB12CD34EF"
          },
          "discount_percent": {
            "type": "number",
            "example": 10
          },
          "discount_amount": {
            "type": "number",
            "example": 125.5
          }
        },
        "required": [
          "code",
          "discount_percent",
          "discount_amount"
        ]
      },
      "OrderResponseDeliveryAddressDto": {
        "type": "object",
        "properties": {
          "city_name": {
            "type": "string",
            "example": "Київ"
          },
          "warehouse_description": {
            "type": "object",
            "example": "Відділення №1: вул. Хрещатик, 1",
            "nullable": true
          },
          "warehouse_number": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "street": {
            "type": "object",
            "example": "вул. Хрещатик",
            "nullable": true
          },
          "building": {
            "type": "object",
            "example": "1А",
            "nullable": true
          },
          "apartment": {
            "type": "object",
            "example": "42",
            "nullable": true
          }
        },
        "required": [
          "city_name",
          "warehouse_description",
          "warehouse_number",
          "street",
          "building",
          "apartment"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "664f1b2c3d4e5f6a7b8c9d10"
          },
          "order_number": {
            "type": "string",
            "example": "FO-0000123"
          },
          "customer": {
            "$ref": "#/components/schemas/OrderResponseCustomerDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponseItemDto"
            }
          },
          "subtotal_price": {
            "type": "number",
            "example": 2598
          },
          "total_price": {
            "type": "number",
            "example": 2338.2
          },
          "applied_discount": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderResponseDiscountDto"
              }
            ]
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "CASH",
              "IBAN",
              "LIQPAY",
              "MONOPAY"
            ]
          },
          "payment_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED",
              "REFUNDED"
            ]
          },
          "payment_transaction_id": {
            "type": "object",
            "example": "liqpay_tx_123",
            "nullable": true
          },
          "delivery_method": {
            "type": "string",
            "enum": [
              "NOVA_POST",
              "COURIER",
              "PICKUP"
            ]
          },
          "delivery_address": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderResponseDeliveryAddressDto"
              }
            ]
          },
          "nova_post_ttn": {
            "type": "object",
            "example": "59001000000000",
            "nullable": true
          },
          "order_status": {
            "type": "string",
            "enum": [
              "NEW",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED",
              "RETURNED"
            ]
          },
          "comment": {
            "type": "object",
            "example": "Зателефонуйте перед відправкою",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "_id",
          "order_number",
          "customer",
          "items",
          "subtotal_price",
          "total_price",
          "applied_discount",
          "payment_method",
          "payment_status",
          "payment_transaction_id",
          "delivery_method",
          "delivery_address",
          "nova_post_ttn",
          "order_status",
          "comment",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 137
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "UpdateOrderCustomerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOrderAddressDto": {
        "type": "object",
        "properties": {}
      },
      "AdminUpdateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          },
          "customer": {
            "$ref": "#/components/schemas/UpdateOrderCustomerDto"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "CASH",
              "IBAN",
              "LIQPAY",
              "MONOPAY"
            ],
            "example": "LIQPAY"
          },
          "delivery_method": {
            "type": "string",
            "enum": [
              "NOVA_POST",
              "COURIER",
              "PICKUP"
            ],
            "example": "NOVA_POST"
          },
          "delivery_address": {
            "$ref": "#/components/schemas/UpdateOrderAddressDto"
          },
          "comment": {
            "type": "string",
            "example": "Зателефонуйте за 30 хв до доставки"
          }
        }
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "order_status": {
            "type": "string",
            "enum": [
              "NEW",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED",
              "RETURNED"
            ],
            "example": "CONFIRMED"
          }
        },
        "required": [
          "order_status"
        ]
      },
      "UpdatePaymentStatusDto": {
        "type": "object",
        "properties": {
          "payment_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED",
              "REFUNDED"
            ],
            "example": "PAID"
          },
          "payment_transaction_id": {
            "type": "string",
            "example": "liqpay_txn_abc123"
          }
        },
        "required": [
          "payment_status"
        ]
      },
      "SetTtnDto": {
        "type": "object",
        "properties": {
          "nova_post_ttn": {
            "type": "string",
            "example": "20450081729182"
          }
        },
        "required": [
          "nova_post_ttn"
        ]
      },
...
    }
  }
}